+++ src/3_game_loop/a_game_logic/entities/key_entity.c
@@ -1,22 +1,22 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   door_entity.c                                      :+:      :+:    :+:   */
+/*   key_entity.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: freddy <freddy@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/06/25 11:57:29 by freddy            #+#    #+#             */
-/*   Updated: 2024/06/25 17:00:16 by freddy           ###   ########.fr       */
+/*   Created: 2024/06/25 15:53:33 by freddy            #+#    #+#             */
+/*   Updated: 2024/06/25 16:10:27 by freddy           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../../include/cub3d.h"
 
-void	tick_door(t_entity *self)
+void	tick_key(t_entity *self)
 {
-	if (pos_distance(player()->transform.pos, self->transform.pos) > DOOR_OPEN_DISTANCE || player()->inv.keys < 1)
+	if (pos_distance(player()->transform.pos, self->transform.pos) > KEY_COLLISION_DISTANCE)
 		return ;
-	logger(LOGGER_INFO, "Door unlocked!");
-	player()->inv.keys--;
-	((t_door *)self->data)->locked = false;
+	logger(LOGGER_INFO, "Key collected!");
+	player()->inv.keys++;
+	delete_entity(self);
 }