+++ src/3_game_loop/a_game_logic/hooks/loop_hook.c
@@ -6,7 +6,7 @@
 /*   By: freddy <freddy@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/28 10:23:28 by freddy            #+#    #+#             */
-/*   Updated: 2024/06/25 15:14:15 by freddy           ###   ########.fr       */
+/*   Updated: 2024/06/25 16:11:50 by freddy           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,6 +25,7 @@ void	loop_hook(void *param)
 	int		prev_frame;
 
 	(void) param;
+	printf("player is currently holding %d keys!\n", player()->inv.keys);
 	curr_time = mlx_get_time();
 	prev_time = game()->prev_time;
 	curr_frame = get_frame(curr_time);